Moved XPathParseTool to fastexternalitemset package
